+++ HISTORY.rst
@@ -33,6 +33,7 @@ Bug tracker at https://github.com/giampaolo/psutil/issues
 - #759: [Linux] Process.memory_maps() may return paths ending with " (deleted)"
 - #761: [Windows] psutil.boot_time() wraps to 0 after 49 days.
 - #764: [NetBSD] fix compilation on NetBSD-6.x.
+- #767: [Linux] disk_io_counters() is broken (ValueError) on Linux 2.4.
 
 
 3.4.2 - 2016-01-20